Flip scroll button based on page position
