+++ src/components/Bio/Bio.scss
