+++ src/hooks/usePrevious.js
@@ -1,0 +1,9 @@
+import { useEffect, useRef } from 'react';
+
+export const usePrevious = (value) => {
+  const ref = useRef();
+  useEffect(() => {
+    ref.current = value;
+  });
+  return ref.current;
+};
